+++ compiler/rustc_parse/src/parser/tokenstream/tests.rs
@@ -1,5 +1,4 @@
-use crate::tests::string_to_stream;
-
+use crate::parser::tests::string_to_stream;
 use rustc_ast::token::{self, IdentIsRaw};
 use rustc_ast::tokenstream::{TokenStream, TokenTree};
 use rustc_span::create_default_session_globals_then;
